Edit: Moved provider factories to infra layer on shared module
