+++ src/modules/selectedGame/infra/repositories/AchievementsRepository.ts
@@ -1,6 +1,6 @@
 import IAchievement from 'modules/selectedGame/entities/IAchievement';
 import IAchievementsRepository from 'modules/selectedGame/repositories/IAchievementsRepository';
-import { makeHttpProvider } from 'shared/container/providers';
+import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
 
 export default class AchievementsRepository implements IAchievementsRepository {
   private httpProvider = makeHttpProvider();
