+++ src/modules/selectedGame/infra/repositories/ActivitiesRepository.ts
@@ -1,6 +1,6 @@
 import IActivity from 'modules/selectedGame/entities/IActivity';
 import IActivitiesRepository from 'modules/selectedGame/repositories/IActivitiesRepository';
-import { makeHttpProvider } from 'shared/container/providers';
+import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
 
 export default class ActivitiesRepository implements IActivitiesRepository {
   private httpProvider = makeHttpProvider();
