+++ src/modules/selectedGame/infra/repositories/FeedPostsRepository.ts
@@ -1,6 +1,6 @@
 import IFeedPost from 'modules/selectedGame/entities/IFeedPost';
 import IFeedPostsRepository from 'modules/selectedGame/repositories/IFeedPostsRepository';
-import { makeHttpProvider } from 'shared/container/providers';
+import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
 
 export default class FeedPostsRepository implements IFeedPostsRepository {
   private httpProvider = makeHttpProvider();
