+++ src/modules/selectedGame/infra/repositories/PlayersRepository.ts
@@ -1,6 +1,6 @@
 import IPlayersRepository from 'modules/selectedGame/repositories/IPlayersRepository';
-import { makeHttpProvider } from 'shared/container/providers';
 import IPlayer from 'shared/domain/entities/IPlayer';
+import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
 
 export default class PlayersRepository implements IPlayersRepository {
   private httpProvider = makeHttpProvider();
