--- src/shared/container/providers/index.ts
@@ -1,3 +1,0 @@
-export { default as makeHttpProvider } from './HTTPProvider';
-export { default as makeJwtProvider } from './JwtProvider';
-export { default as makeStorageProvider } from './StorageProvider';
